pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "pillow>=11.3.0",
    "streamlit>=1.47.1",
]
